Added all test cases for the angle of triangle function

+++ PolygonChecker/TriangleInsideAngle.h
@@ -1,3 +1,3 @@
 #pragma once
 
-double anglesofTriangle(double side1, double side2, double side3); 
+int anglesofTriangle(double side1, double side2, double side3);
